+++ src/util/mod.rs
@@ -43,6 +43,7 @@ pub mod proxy;
 pub mod raw_archive;
 pub mod response_cache;
 pub mod scan;
+pub mod secret_link;
 pub mod see_know;
 pub mod service_defs;
 pub mod settings;